requirements.txt
torch>=2.0.0
sounddevice>=0.4.6
google-genai>=0.2.0
omegaconf>=2.3.0